Some more renaming: command/commit/commit.py -> command/commit.py
