+++ src/mepo/command/fetch.py
@@ -1,15 +1,16 @@
-from mepo.state.state import MepoState
-from mepo.utilities import verify
-from mepo.repository.git import GitRepository
+from ..state.state import MepoState
+from ..utilities import colors
+from ..utilities import verify
+from ..repository.git import GitRepository
 
 def run(args):
     allcomps = MepoState.read_state()
-    verify.valid_components(args.comp_name, allcomps)
-    comps2deltg = _get_comps_to_list(args.comp_name, allcomps)
-    for comp in comps2deltg:
+    comps2fetch = _get_comps_to_list(args.comp_name, allcomps)
+    for comp in comps2fetch:
         git = GitRepository(comp.remote, comp.local)
-        git.delete_tag(args.tag_name)
-        print('- {}: {}'.format(comp.name, args.tag_name))
+        print("Fetching %s" %
+                 colors.YELLOW + comp.name + colors.RESET)
+        git.fetch(args)
 
 def _get_comps_to_list(specified_comps, allcomps):
     comps_to_list = allcomps
