+++ src/mepo/command/restore-state.py
@@ -3,10 +3,10 @@ import time
 import multiprocessing as mp
 import atexit
 
-from mepo.state.state import MepoState
-from mepo.repository.git import GitRepository
-from mepo.utilities.version import version_to_string
-from mepo.utilities import colors
+from ..state.state import MepoState
+from ..repository.git import GitRepository
+from ..utilities.version import version_to_string
+from ..utilities import colors
 
 def run(args):
     print('Checking status...'); sys.stdout.flush()
