+++ src/mepo/command/tag_delete.py
@@ -1,16 +1,15 @@
-from mepo.state.state import MepoState
-from mepo.utilities import verify
-from mepo.repository.git import GitRepository
-from mepo.utilities import colors
+from ..state.state import MepoState
+from ..utilities import verify
+from ..repository.git import GitRepository
 
 def run(args):
     allcomps = MepoState.read_state()
-    comps2fetch = _get_comps_to_list(args.comp_name, allcomps)
-    for comp in comps2fetch:
+    verify.valid_components(args.comp_name, allcomps)
+    comps2deltg = _get_comps_to_list(args.comp_name, allcomps)
+    for comp in comps2deltg:
         git = GitRepository(comp.remote, comp.local)
-        print("Fetching %s" %
-                 colors.YELLOW + comp.name + colors.RESET)
-        git.fetch(args)
+        git.delete_tag(args.tag_name)
+        print('- {}: {}'.format(comp.name, args.tag_name))
 
 def _get_comps_to_list(specified_comps, allcomps):
     comps_to_list = allcomps
