+++ src/mepo/command/tag_push.py
@@ -1,6 +1,6 @@
-from mepo.state.state import MepoState
-from mepo.utilities import verify
-from mepo.repository.git import GitRepository
+from ..state.state import MepoState
+from ..utilities import verify
+from ..repository.git import GitRepository
 
 def run(args):
     allcomps = MepoState.read_state()
